requirements.txt
python-dotenv
py-cord[voice]
psycopg2-binary
requests